pyproject.toml
[tool.poetry]
name = "ctbot"
version = "0.1.0"
description = "A Simplified Trading Bot for Binance USDT-M Futures Testnet"
authors = ["Gemini"]
packages = [{include = "bot", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
python-binance = "^1.0.19"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.0"
rich = "^13.6.0"
flask = "^3.0.0"
flask-cors = "^4.0.0"
sqlalchemy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
